GOF_Design_Patterns/test/com/test/gof/design/patterns/creational/SingletonTest.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package com.test.gof.design.patterns.creational;

import com.gof.design.patterns.creational.singleton.Singleton;
import java.util.Date;
import static org.testng.Assert.*;
import org.testng.annotations.AfterClass;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

/**
 *
 * @author Azariahs
 */
public class SingletonTest {
    
    public SingletonTest() {
    }
    // TODO add test methods here.
    // The methods must be annotated with annotation @Test. For example:
    //
    // @Test
    // public void hello() {}

    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @BeforeMethod
    public void setUpMethod() throws Exception {
    }

    @AfterMethod
    public void tearDownMethod() throws Exception {
    }
    
    @Test
    public void testForSingleton() {
        Singleton a = Singleton.instance();
        Singleton b = Singleton.instance();
        
        assertEquals(a, b);
        
        Date date1 = a.getDate();
        Date date2 = b.getDate();
        
        assertEquals(date1, date2);
    }
}